refactor to node test runner
