+++ .eslintrc.json
@@ -1,17 +1,20 @@
 {
   "extends": [
     "eslint:recommended",
-    "plugin:node/recommended",
     "plugin:react/recommended",
     "plugin:react-hooks/recommended"
   ],
   "parserOptions": {
-    "ecmaVersion": "latest"
+    "ecmaVersion": "latest",
+    "sourceType": "script"
   },
   "settings": {
     "react": {
       "version": "detect"
     }
+  },
+  "env": {
+    "node": true
   },
   "rules": {
     "arrow-body-style": ["error", "always"],
