+++ package.json
@@ -51,20 +51,18 @@
     "test": "npm run test:lint && npm run test:prettier && npm run test:unit",
     "test:lint": "eslint .",
     "test:prettier": "prettier -c .",
-    "test:unit": "node test/run.mjs",
+    "test:unit": "node --test",
     "prepublishOnly": "npm test"
   },
   "devDependencies": {
     "eslint": "^8.8.0",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-react": "^7.23.2",
     "eslint-plugin-react-hooks": "^4.2.0",
     "prettier": "^2.3.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-test-renderer": "^18.2.0",
-    "test-director": "^10.0.0"
+    "react-test-renderer": "^18.2.0"
   },
   "dependencies": {
     "@emotion/react": "^11.4.0",
